pyproject.toml
[project]
name = "weather-api"
version = "0.1.0"
description = "Coleta OpenWeather e expõe via FastAPI"
authors = [
    {name = "Ricardo Sprocati",email = "ricardosprocatioliveira@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0)",
    "alembic (>=1.13)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "pydantic-settings (>=2.6)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
